odk-api/requirements.txt
aio-pika==6.6.1
aiormq==3.2.3
appdirs==1.4.4
astroid==2.4.2
attrs==19.3.0
black==19.10b0
cfgv==3.2.0
click==7.1.2
databases==0.3.2
distlib==0.3.1
fastapi==0.61.0
filelock==3.0.12
gunicorn==20.0.4
h11==0.9.0
httptools==0.1.1
identify==1.4.25
idna==2.10
isort==4.3.21
lazy-object-proxy==1.5.1
loguru==0.5.1
mccabe==0.6.1
more-itertools==8.4.0
multidict==4.7.6
mypy==0.782
mypy-extensions==0.4.3
nodeenv==1.4.0
packaging==20.4
pamqp==2.3.0
pathspec==0.8.0
pip-review==1.1.0
pipupgrade==1.7.4
pkg-resources==0.0.0
pluggy==0.13.1
pre-commit==2.6.0
psycopg2-binary==2.8.5
py==1.9.0
pycodestyle==2.6.0
pydantic==1.6.1
pylint==2.5.3
pyparsing==2.4.7
pytest==5.4.3
PyYAML==5.3.1
regex==2020.7.14
rope==0.17.0
six==1.15.0
SQLAlchemy==1.3.18
starlette==0.13.8
tenacity==6.2.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
uvicorn==0.11.8
uvloop==0.14.0
virtualenv==20.0.30
wcwidth==0.2.5
websockets==8.1
wrapt==1.12.1
yarl==1.5.1
